added my zine
slideshow not entirely working. have to click the next button 6 times for an image to show up. copy pasted from jacksons, not sure if that will cause any issues with on his. thinking about reformatting images with borders so they are all the same shape, that way it wont resize the page when they click thru